refactor compile routine
- split compile.ts to multiple helper files
- use draft-07 from ajv
